+++ tests/Pfarah.Tests/BinaryTests.fs
@@ -52,7 +52,7 @@ let ``load binary data`` () =
   let lookup = dict([(0x284ds, "date")])
   let stream = strm([|0x45; 0x55; 0x34; 0x62; 0x69; 0x6e; 0x4d; 0x28;
                       0x01; 0x00; 0x0c; 0x00; 0x10; 0x77; 0x5d; 0x03|])
-  ParaValue.LoadWithHeader(stream, "EU4bin", "EU4txt", lazy lookup)
+  ParaValue.Load(stream, "EU4bin", "EU4txt", lazy lookup)
   |> shouldEqual (ParaValue.Record([| ("date", ParaValue.Date(DateTime(1444, 11, 11))) |]))
 
 [<Test>]
@@ -368,11 +368,11 @@ let ``binary parse heterogeneous array`` () =
 [<Test>]
 let ``load plain text file`` () =
   let path = Path.Combine("data", "eu4bin.eu4")
-  ParaValue.LoadFile(path, "EU4bin", "EU4txt", lazy(dict([(0x284ds, "date")])))
+  ParaValue.Load(path, "EU4bin", "EU4txt", lazy(dict([(0x284ds, "date")])))
   |> shouldEqual (ParaValue.Record([| ("date", ParaValue.Date (DateTime(1757, 8, 12)))|]))
 
 [<Test>]
 let ``load zip text file`` () =
   let path = Path.Combine("data", "eu4bin-zip.eu4")
-  ParaValue.LoadFile(path, "EU4bin", "EU4txt", lazy(dict([(0x284ds, "date")])))
+  ParaValue.Load(path, "EU4bin", "EU4txt", lazy(dict([(0x284ds, "date")])))
   |> shouldEqual (ParaValue.Record([| ("date", ParaValue.Date (DateTime(1757, 8, 12)))|]))